requirements.txt
Flask==3.1.0
tensorflow==2.15.0
apify-client==1.9.4
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
pillow==10.3.0
requests==2.31.0
gunicorn==22.0.0
